Tests packages refactored.
